requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/base.in
#
altgraph==0.17.3
    # via
    #   macholib
    #   modulegraph
    #   py2app
macholib==1.16.2
    # via py2app
modulegraph==0.19.3
    # via py2app
py2app==0.28.4
    # via -r requirements/base.in
pyobjc-core==9.0.1
    # via pyobjc-framework-cocoa
pyobjc-framework-cocoa==9.0.1
    # via rumps
pyperclip==1.8.2
    # via -r requirements/base.in
python-gnupg==0.5.0
    # via -r requirements/base.in
rumps==0.4.0
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
